logging unification between twisted plugin and pyrdp-mitm.py
Refactored common logging code out in pyrdp.mitm.cli
