+++ pyrdp/mitm/cli.py
@@ -10,14 +10,18 @@ To be consumed either via bin/pyrdp-mitm.py or via twistd plugin.
 """
 import argparse
 import logging
+import logging.handlers
 import os
 import sys
+from pathlib import Path
 from typing import Tuple
 
 import appdirs
 import OpenSSL
 
 from pyrdp.core.ssl import ServerTLSContext
+from pyrdp.logging import JSONFormatter, log, LOGGER_NAMES, LoggerNameFilter, SessionLogger, VariableFormatter
+from pyrdp.mitm.config import MITMConfig
 
 
 def parseTarget(target: str) -> Tuple[str, int]:
@@ -57,6 +61,7 @@ def validateKeyAndCertificate(private_key: str, certificate: str) -> Tuple[str, 
         sys.exit(1)
 
     return key, cert
+
 
 def handleKeyAndCertificate(key: str, certificate: str):
     """
@@ -111,3 +116,62 @@ def generateCertificate(keyPath: str, certificatePath: str) -> bool:
 
     result = os.system("openssl req -newkey rsa:2048 -nodes -keyout %s -x509 -days 365 -out %s -subj \"/CN=www.example.com/O=PYRDP/C=US\" 2>%s" % (keyPath, certificatePath, nullDevicePath))
     return result == 0
+
+
+def prepareLoggers(logLevel: int, logFilter: str, sensorID: str, outDir: Path):
+    """
+    :param logLevel: log level for the stream handler.
+    :param logFilter: logger name to filter on.
+    :param sensorID: ID to differentiate between instances of this program in the JSON log.
+    :param outDir: output directory.
+    """
+    logDir = outDir / "logs"
+    logDir.mkdir(exist_ok = True)
+
+    formatter = VariableFormatter("[{asctime}] - {levelname} - {sessionID} - {name} - {message}", style = "{", defaultVariables = {
+        "sessionID": "GLOBAL"
+    })
+
+    streamHandler = logging.StreamHandler()
+    streamHandler.setFormatter(formatter)
+    streamHandler.setLevel(logLevel)
+    streamHandler.addFilter(LoggerNameFilter(logFilter))
+
+    logFileHandler = logging.handlers.TimedRotatingFileHandler(logDir / "mitm.log", when = "D")
+    logFileHandler.setFormatter(formatter)
+
+    jsonFileHandler = logging.FileHandler(logDir / "mitm.json")
+    jsonFileHandler.setFormatter(JSONFormatter({"sensor": sensorID}))
+    jsonFileHandler.setLevel(logging.INFO)
+
+    rootLogger = logging.getLogger(LOGGER_NAMES.PYRDP)
+    rootLogger.addHandler(streamHandler)
+    rootLogger.addHandler(logFileHandler)
+    rootLogger.setLevel(logging.DEBUG)
+
+    connectionsLogger = logging.getLogger(LOGGER_NAMES.MITM_CONNECTIONS)
+    connectionsLogger.addHandler(jsonFileHandler)
+
+    crawlerFormatter = VariableFormatter("[{asctime}] - {sessionID} - {message}", style = "{", defaultVariables = {
+        "sessionID": "GLOBAL"
+    })
+
+    crawlerFileHandler = logging.FileHandler(logDir / "crawl.log")
+    crawlerFileHandler.setFormatter(crawlerFormatter)
+
+    jsonCrawlerFileHandler = logging.FileHandler(logDir / "crawl.json")
+    jsonCrawlerFileHandler.setFormatter(JSONFormatter({"sensor": sensorID}))
+
+    crawlerLogger = logging.getLogger(LOGGER_NAMES.CRAWLER)
+    crawlerLogger.addHandler(crawlerFileHandler)
+    crawlerLogger.addHandler(jsonCrawlerFileHandler)
+    crawlerLogger.setLevel(logging.INFO)
+
+    log.prepareSSLLogger(logDir / "ssl.log")
+
+    return logging.getLogger(LOGGER_NAMES.MITM)
+
+
+def logConfiguration(config: MITMConfig):
+    logging.getLogger(LOGGER_NAMES.MITM).info("Target: %(target)s:%(port)d", {"target": config.targetHost, "port": config.targetPort})
+    logging.getLogger(LOGGER_NAMES.MITM).info("Output directory: %(outputDirectory)s", {"outputDirectory": config.outDir.absolute()})
